refactor : 타입 하나로 통일
