+++ src/hooks/useScheduleForm.ts
@@ -1,13 +1,13 @@
 /* eslint-disable array-callback-return */
 import { postSchedulesAPI } from 'libs/api/schedule';
-import { Schedule } from 'types/schedule';
+import { Schedule } from 'types/customTypes';
 
 const useScheduleForm = () => {
   const onSubmitSchedule = async (weekdays: string[], body: Schedule) => {
     await Promise.all(
       weekdays.map((weekday) => {
         postSchedulesAPI(weekday, {
-          id: Math.floor(Math.random() * 10000),
+          id: Math.floor(Math.random() * 10000).toString(),
           startTime: body.startTime,
           endTime: body.endTime,
         });
